Add license_helper.sh
License_helper is a simple script that makes sure every file contains a
license header.
It takes the current year and and uses the name configured as the
current git user.

In the long run I would like to improve on this by getting contributers
and years a file changed in from git and making sure the license headers
are always up to date in the script.