Add View and fancy output hook
Added 'view', 'greedyView' and 'copyView', where the first 2 are from
xmonad, the third enables to map multiple outputs to the same workspace.

Also added the "fancy" output add hook, which only maps a workspace to
the output automatically if waymonad hasn't been running for at least 5
minutes.
This is the "presentation protection". Multi-Monitor setups will be
added quickly enough, but plugging in a projector or similar will not
show any windows by default